# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OM PAY API",
        "description": "API de paiement mobile OM PAY avec authentification OTP et gestion de comptes",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Serveur de développement"
        },
        {
            "url": "https://ompay-isuf.onrender.com/api",
            "description": "Serveur de production"
        }
    ],
    "paths": {
...
        "/transactions/paiement": {
            "post": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Effectuer un paiement",
                "description": "Effectue un paiement vers un marchand (code marchand) ou vers un client (numéro de téléphone)",
                "operationId": "paiement",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "destinataire",
                                    "montant"
                                ],
                                "properties": {
                                    "destinataire": {
                                        "description": "Code marchand ou numéro de téléphone du destinataire",
                                        "type": "string",
                                        "example": "MARCHAND001"
                                    },
                                    "montant": {
                                        "description": "Montant du paiement",
                                        "type": "number",
                                        "format": "float",
                                        "example": 2500
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Paiement effectué avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Paiement effectué avec succès"
                                        },
                                        "data": {
                                            "properties": {
                                                "transaction": {
                                                    "properties": {
                                                        "reference": {
                                                            "type": "string",
                                                            "example": "TXN-20251110-ABC123"
                                                        },
                                                        "type": {
                                                            "type": "string",
                                                            "example": "paiement"
                                                        },
                                                        "montant": {
                                                            "type": "number",
                                                            "format": "float",
                                                            "example": 2500
                                                        },
                                                        "frais": {
                                                            "type": "number",
                                                            "format": "float",
                                                            "example": 0
                                                        },
                                                        "statut": {
                                                            "type": "string",
                                                            "example": "reussi"
                                                        },
                                                        "compteEmetteur": {
                                                            "properties": {
                                                                "numero_compte": {
                                                                    "type": "string",
                                                                    "example": "OM-2025-AB12-CD34"
                                                                }
                                                            },
                                                            "type": "object"
                                                        },
                                                        "marchand": {
                                                            "properties": {
                                                                "raison_sociale": {
                                                                    "type": "string",
                                                                    "example": "Boutique Express"
                                                                },
                                                                "code_marchand": {
                                                                    "type": "string",
                                                                    "example": "MARCHAND001"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erreur de traitement (solde insuffisant, montant invalide, etc.)"
                    },
                    "404": {
                        "description": "Marchand ou compte non trouvé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/transactions": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Liste des transactions de l'utilisateur",
                "description": "Retourne la liste paginée des transactions de l'utilisateur connecté (triées par date décroissante)",
                "operationId": "getTransactions",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numéro de la page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Nombre d'éléments par page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15,
                            "maximum": 100,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des transactions paginée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "transactions": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "reference": {
                                                                "type": "string",
                                                                "example": "TXN-20251110-ABC123"
                                                            },
                                                            "type": {
                                                                "type": "string",
                                                                "enum": [
                                                                    "paiement",
                                                                    "transfert",
                                                                    "depot"
                                                                ],
                                                                "example": "transfert"
                                                            },
                                                            "montant": {
                                                                "type": "number",
                                                                "format": "float",
                                                                "example": 5000
                                                            },
                                                            "frais": {
                                                                "type": "number",
                                                                "format": "float",
                                                                "example": 50
                                                            },
                                                            "statut": {
                                                                "type": "string",
                                                                "enum": [
                                                                    "reussi",
                                                                    "echec"
                                                                ],
                                                                "example": "reussi"
                                                            },
                                                            "created_at": {
                                                                "type": "string",
                                                                "format": "date-time",
                                                                "example": "2025-11-10T09:00:00Z"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "pagination": {
                                                    "properties": {
                                                        "current_page": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "last_page": {
                                                            "type": "integer",
                                                            "example": 5
                                                        },
                                                        "per_page": {
                                                            "type": "integer",
                                                            "example": 15
                                                        },
                                                        "total": {
                                                            "type": "integer",
                                                            "example": 67
                                                        },
                                                        "from": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "to": {
                                                            "type": "integer",
                                                            "example": 15
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}